+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### next release
 - :bulb: Added `Api::V1::SubscriptionsController` for subscribe/unsubscribe actions
+- :bulb: Add show information action for `Api::V1::ContactsController`
+- :bulb: Add show information action for `Api::V1::NotificationsController`
 
 - :bulb: Added list with all tracking events
 - :bulb: Added feature to track emails opening