+++ CHANGELOG.md
@@ -1,14 +1,13 @@
 # Changelog
 
-### next release
+### v0.7.0
 - :bulb: Added `Api::V1::SubscriptionsController` for subscribe/unsubscribe actions
 - :bulb: Add show information action for `Api::V1::ContactsController`
 - :bulb: Add show information action for `Api::V1::NotificationsController`
-- :arrow_up: Updated API documentation
-
 - :bulb: Added list with all tracking events
 - :bulb: Added feature to track emails opening
 - :bulb: Implemented events dispatching
+- :arrow_up: Updated API documentation
 
 ### v0.6.6
 - :bulb: Added feature to clean all contacts statuses